[Fix #159965441] Remove extra nav bar in the new article page
